test: move out git_utils fto test_utils crate
