+++ Cargo.toml
@@ -1,3 +1,5 @@
+workspace = { members = ["test-utils"] }
+
 [package]
 name = "nextsv"
 version = "0.7.9"
@@ -10,13 +12,13 @@ readme = "README.md"
 repository = "https://github.com/jerusdp/nextsv"
 keywords = ["semantic", "version", "calculator"]
 include = ["**/*.rs", "Cargo.toml", "README.md", "LICENSE", "CHANGES.md"]
+publish = true
 
 
 [dependencies]
 clap = { version = "4.0.30", features = ["derive"] }
 clap-verbosity-flag = "2.0.0"
 colored = "2.1.0"
-copy_dir = "0.1.3"
 env_logger = "0.11.1"
 git-conventional = "0.12.0"
 git2 = "0.18.1"
@@ -27,16 +29,19 @@ regex = "1.10.4"
 thiserror = "1.0"
 
 [dev-dependencies]
-chrono = "0.4.38"
+# chrono = "0.4.38"
+# copy_dir = "0.1.3"
+snapbox = { version = "0.5.9", features = ["cmd"] }
+# uuid = { version = "1.8.0", features = ["v4"] }
 claims = "0.7"
 log4rs_test_utils = { version = "0.2.3", default-features = false, features = [
     "test_logging",
 ] }
 map-macro = "0.3.0"
-rstest = "0.18.2"
-snapbox = { version = "0.5.9", features = ["cmd"] }
+rstest = "0.19.0"
 trycmd = "0.15.0"
-uuid = { version = "1.8.0", features = ["v4"] }
+test_utils = { path = "test-utils" }
+
 
 [lib]
 name = "nextsv"
