+++ test-utils/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+git2 = "0.18.3"
+snapbox = "0.5.9"
+uuid = "1.8.0"
+copy_dir = "0.1.3"
+chrono = "0.4.38"
